+++ HousingSearchListener.Tests/V1/E2ETests/Stories/AddPersonToIndexTests.cs
@@ -10,23 +10,24 @@ namespace HousingSearchListener.Tests.V1.E2ETests.Stories
 {
     [Story(
         AsA = "SQS Tenure Listener",
-        IWant = "a function to process the Asset created and updated messages",
-        SoThat = "The Asset details are set in the index")]
+        IWant = "a function to process the person created and updated messages",
+        SoThat = "The person details are set in the index")]
     [Collection("ElasticSearch collection")]
-    public class AddAssetToIndexTests : IDisposable
+    public class AddPersonToIndexTests : IDisposable
     {
         private readonly ElasticSearchFixture _esFixture;
-        private readonly AssetApiFixture _AssetApiFixture;
+        private readonly PersonApiFixture _personApiFixture;
         private readonly TenureApiFixture _tenureApiFixture;
 
-        private readonly AddAssetToIndexSteps _steps;
+        private readonly AddPersonToIndexSteps _steps;
 
-        public AddAssetToIndexTests(ElasticSearchFixture esFixture)
+        public AddPersonToIndexTests(ElasticSearchFixture esFixture)
         {
             _esFixture = esFixture;
-            _AssetApiFixture = new AssetApiFixture();
+            _personApiFixture = new PersonApiFixture();
+            _tenureApiFixture = new TenureApiFixture();
 
-            _steps = new AddAssetToIndexSteps();
+            _steps = new AddPersonToIndexSteps();
         }
 
         public void Dispose()
@@ -40,7 +41,7 @@ namespace HousingSearchListener.Tests.V1.E2ETests.Stories
         {
             if (disposing && !_disposed)
             {
-                _AssetApiFixture.Dispose();
+                _personApiFixture.Dispose();
                 _tenureApiFixture.Dispose();
 
                 _disposed = true;
@@ -48,25 +49,54 @@ namespace HousingSearchListener.Tests.V1.E2ETests.Stories
         }
 
         [Theory]
-        [InlineData(EventTypes.AssetCreatedEvent)]
-        public void AssetNotFound(string eventType)
+        [InlineData(EventTypes.PersonCreatedEvent)]
+        [InlineData(EventTypes.PersonUpdatedEvent)]
+        public void PersonNotFound(string eventType)
         {
-            var AssetId = Guid.NewGuid();
-            this.Given(g => _AssetApiFixture.GivenTheAssetDoesNotExist(AssetId))
-                .When(w => _steps.WhenTheFunctionIsTriggered(AssetId, eventType))
-                .Then(t => _steps.ThenTheCorrelationIdWasUsedInTheApiCall(_AssetApiFixture.ReceivedCorrelationIds))
-                .Then(t => _steps.ThenAAssetNotFoundExceptionIsThrown(AssetId))
+            var personId = Guid.NewGuid();
+            this.Given(g => _personApiFixture.GivenThePersonDoesNotExist(personId))
+                .When(w => _steps.WhenTheFunctionIsTriggered(personId, eventType))
+                .Then(t => _steps.ThenTheCorrelationIdWasUsedInTheApiCall(_personApiFixture.ReceivedCorrelationIds))
+                .Then(t => _steps.ThenAPersonNotFoundExceptionIsThrown(personId))
                 .BDDfy();
         }
 
         [Fact]
-        public void AssetCreatedAddedToIndex()
+        public void PersonCreatedAddedToIndex()
         {
-            var AssetId = Guid.NewGuid();
-            this.Given(g => _AssetApiFixture.GivenTheAssetExists(AssetId))
-                .When(w => _steps.WhenTheFunctionIsTriggered(AssetId, EventTypes.AssetCreatedEvent))
-                .Then(t => _steps.ThenTheCorrelationIdWasUsedInTheApiCall(_AssetApiFixture.ReceivedCorrelationIds))
-                .Then(t => _steps.ThenTheIndexIsUpdatedWithTheAsset(_AssetApiFixture.ResponseObject, _esFixture.ElasticSearchClient))
+            var personId = Guid.NewGuid();
+            this.Given(g => _personApiFixture.GivenThePersonExists(personId))
+                .And(h => _esFixture.GivenAPersonIsNotIndexed(_personApiFixture.ResponseObject))
+                .When(w => _steps.WhenTheFunctionIsTriggered(personId, EventTypes.PersonCreatedEvent))
+                .Then(t => _steps.ThenTheCorrelationIdWasUsedInTheApiCall(_personApiFixture.ReceivedCorrelationIds))
+                .Then(t => _steps.ThenTheIndexIsUpdatedWithThePerson(_personApiFixture.ResponseObject, _esFixture.ElasticSearchClient))
+                .BDDfy();
+        }
+
+        [Fact]
+        public void PersonUpdateInIndex()
+        {
+            var personId = Guid.NewGuid();
+            this.Given(g => _personApiFixture.GivenThePersonExists(personId))
+                .And(h => _esFixture.GivenAPersonIsIndexedWithDifferentInfo(_personApiFixture.ResponseObject))
+                .When(w => _steps.WhenTheFunctionIsTriggered(personId, EventTypes.PersonUpdatedEvent))
+                .Then(t => _steps.ThenTheCorrelationIdWasUsedInTheApiCall(_personApiFixture.ReceivedCorrelationIds))
+                .Then(t => _steps.ThenTheIndexIsUpdatedWithThePerson(_personApiFixture.ResponseObject, _esFixture.ElasticSearchClient))
+                .BDDfy();
+        }
+
+        [Fact]
+        public void TenureUpdateInIndex()
+        {
+            var personId = Guid.NewGuid();
+            this.Given(g => _personApiFixture.GivenThePersonExists(personId))
+                .And(h => _tenureApiFixture.GivenTheTenureExists(
+                                                Guid.Parse(_personApiFixture.ResponseObject.Tenures.First().Id), personId))
+                .And(h => _esFixture.GivenAPersonIsIndexedWithDifferentInfo(_personApiFixture.ResponseObject))
+                .When(w => _steps.WhenTheFunctionIsTriggered(personId, EventTypes.PersonUpdatedEvent))
+                .Then(t => _steps.ThenTheIndexIsUpdatedWithTheUpdatedPersonTenure(_personApiFixture.ResponseObject, _esFixture.ElasticSearchClient))
+                .Then(t => _steps.ThenTheCorrelationIdWasUsedInTheApiCall(_personApiFixture.ReceivedCorrelationIds))
+                .Then(t => _steps.ThenTheCorrelationIdWasUsedInTheApiCall(_tenureApiFixture.ReceivedCorrelationIds))
                 .BDDfy();
         }
     }
